mlogger 0.21

+++ deployable/src/main/java/com/jivesoftware/os/routing/bird/deployable/Deployable.java
@@ -451,7 +451,7 @@ public class Deployable {
 
         @Override
         public void run() {
-            long errors = loggerSummary.errors;
+            long errors = loggerSummary.errors.longValue();
             long delta = errors - lastErrorCount.getAndSet(errors);
             lastErrorDelta.set(delta);
             if (delta > 0) {
